Begin scanner refactor
